requirements.txt
pandas
numpy
Faker
mysql-connector-python
scikit-learn
matplotlib
seaborn
